+++ app/src/main/java/com/tomlezmy/goolmathapp/interfaces/MyDialogListener.java
@@ -1,5 +1,5 @@
 package com.tomlezmy.goolmathapp.interfaces;
 
 public interface MyDialogListener {
-    void onReturn(int answer);
+    void onReturn(float answer);
 }
